Don't set ortho slice to (0,0,0) when no transform
When file does not specify origin or/and voxel size then transform matrix is
identity. However when loading such file we try to set initial slice position
to (0,0,0) regardless there is no coordinate system given. In such case we land
in the top left corner, which is a bit confusing. This change checks if
transform is identify and then keeps initial slice position at the center of
image space.

This makes MRIcroGL also in par with MacOS MRIcro.